chore: rename webhook name
