+++ cmd/cosign/webhook/main.go
@@ -45,14 +45,14 @@ import (
 var secretName = flag.String("secret-name", "", "The name of the secret in the webhook's namespace that holds the public key for verification.")
 
 // webhookName holds the name of the validating and mutating webhook
-// configuration resources dispatching admission requests to cosigned.
+// configuration resources dispatching admission requests to policy-controller.
 // It is also the name of the webhook which is injected by the controller
 // with the resource types, namespace selectors, CABindle and service path.
 // If this changes, you must also change:
 //    ./config/500-webhook-configuration.yaml
-//    https://github.com/sigstore/helm-charts/blob/main/charts/cosigned/templates/webhook/webhook_mutating.yaml
-//    https://github.com/sigstore/helm-charts/blob/main/charts/cosigned/templates/webhook/webhook_validating.yaml
-var webhookName = flag.String("webhook-name", "cosigned.sigstore.dev", "The name of the validating and mutating webhook configurations as well as the webhook name that is automatically configured, if exists, with different rules and client settings setting how the admission requests to be dispatched to cosigned.")
+//    https://github.com/sigstore/helm-charts/blob/main/charts/policy-controller/templates/webhook/webhook_mutating.yaml
+//    https://github.com/sigstore/helm-charts/blob/main/charts/policy-controller/templates/webhook/webhook_validating.yaml
+var webhookName = flag.String("webhook-name", "policycontroller.sigstore.dev", "The name of the validating and mutating webhook configurations as well as the webhook name that is automatically configured, if exists, with different rules and client settings setting how the admission requests to be dispatched to policy-controller.")
 
 func main() {
 	opts := webhook.Options{
@@ -69,7 +69,7 @@ func main() {
 	vJSON, _ := v.JSONString()
 	log.Printf("%v", vJSON)
 	// This calls flag.Parse()
-	sharedmain.MainWithContext(ctx, "cosigned",
+	sharedmain.MainWithContext(ctx, "policy-controller",
 		certificates.NewController,
 		NewValidatingAdmissionController,
 		NewMutatingAdmissionController,
